Removing free-form formatting due to resulting parsing reasons
This removal is entirely due to the fact that visibility modifiers could
also be a binary expression operator. This fact, combined with the
free-form formatting, resulted in the parser not being able to discern
between the end of an expression on one line and the start of a
'private' statement declaration on another. As a result, the free-from
formatting has been disabled/removed in order to address this issue.